Moved Matches to Decoder
