+++ pkg/model/v1/decoder.go
@@ -2,6 +2,7 @@ package v1
 
 import (
 	"github.com/gogo/protobuf/proto"
+	"github.com/grafana/tempo/pkg/model/tracepb"
 	"github.com/grafana/tempo/pkg/tempopb"
 )
 
@@ -37,8 +38,14 @@ func (d *Decoder) Unmarshal(obj []byte) (*tempopb.Trace, error) {
 }
 
 func (d *Decoder) Matches(id []byte, obj []byte, req *tempopb.SearchRequest) (*tempopb.TraceSearchMetadata, error) {
-	return nil, nil
+	trace, err := d.Unmarshal(obj)
+	if err != nil {
+		return nil, err
+	}
+
+	return tracepb.MatchesProto(id, trace, req)
 }
+
 func (d *Decoder) Range(obj []byte) (uint32, uint32, error) {
-	return 0, 0, nil
+	return 0, 0, nil // jpe unsupported
 }
